add scenarios, create base view class
